+++ flag/service/operatorkit/operatorkit.go
@@ -1,0 +1,7 @@
+package operatorkit
+
+// OperatorKit is a data structure to hold operatorkit specific
+// configuration.
+type OperatorKit struct {
+	RemovedFinalizersCacheTTL string
+}