+++ main.go
@@ -159,6 +159,7 @@ func mainWithError() (err error) {
 	daemonCommand.PersistentFlags().String(f.Service.Kubernetes.TLS.CrtFile, "", "Certificate file path to use to authenticate with Kubernetes.")
 	daemonCommand.PersistentFlags().String(f.Service.Kubernetes.TLS.KeyFile, "", "Key file path to use to authenticate with Kubernetes.")
 	daemonCommand.PersistentFlags().String(f.Service.Kubernetes.Watch.Namespace, "default", "The namespace where appcatalog and app CRs are located.")
+	daemonCommand.PersistentFlags().String(f.Service.OperatorKit.RemovedFinalizersCacheTTL, "15m", "TTL for cache of removed finalizers.")
 
 	err = newCommand.CobraCommand().Execute()
 	if err != nil {