+++ flag/service/helm/helm.go
@@ -1,7 +1,7 @@
 package helm
 
 import (
-	"github.com/giantswarm/app-operator/v2/flag/service/helm/http"
+	"github.com/giantswarm/app-operator/v3/flag/service/helm/http"
 )
 
 type Helm struct {