+++ flag/service/service.go
@@ -3,10 +3,10 @@ package service
 import (
 	"github.com/giantswarm/operatorkit/v4/pkg/flag/service/kubernetes"
 
-	"github.com/giantswarm/app-operator/v2/flag/service/app"
-	"github.com/giantswarm/app-operator/v2/flag/service/chart"
-	"github.com/giantswarm/app-operator/v2/flag/service/helm"
-	"github.com/giantswarm/app-operator/v2/flag/service/image"
+	"github.com/giantswarm/app-operator/v3/flag/service/app"
+	"github.com/giantswarm/app-operator/v3/flag/service/chart"
+	"github.com/giantswarm/app-operator/v3/flag/service/helm"
+	"github.com/giantswarm/app-operator/v3/flag/service/image"
 )
 
 // Service is an intermediate data structure for command line configuration flags.