+++ flag/service/service.go
@@ -7,13 +7,15 @@ import (
 	"github.com/giantswarm/app-operator/v2/flag/service/chart"
 	"github.com/giantswarm/app-operator/v2/flag/service/helm"
 	"github.com/giantswarm/app-operator/v2/flag/service/image"
+	"github.com/giantswarm/app-operator/v2/flag/service/operatorkit"
 )
 
 // Service is an intermediate data structure for command line configuration flags.
 type Service struct {
-	App        app.App
-	Chart      chart.Chart
-	Helm       helm.Helm
-	Image      image.Image
-	Kubernetes kubernetes.Kubernetes
+	App         app.App
+	Chart       chart.Chart
+	Helm        helm.Helm
+	Image       image.Image
+	Kubernetes  kubernetes.Kubernetes
+	Operatorkit operatorkit.Operatorkit
 }