+++ service/controller/app/app.go
@@ -26,6 +26,7 @@ type Config struct {
 	ChartNamespace    string
 	HTTPClientTimeout time.Duration
 	ImageRegistry     string
+	ResyncPeriod      time.Duration
 	UniqueApp         bool
 	WebhookAuthToken  string
 	WebhookBaseURL    string
@@ -53,6 +54,9 @@ func NewApp(config Config) (*App, error) {
 	}
 	if config.ImageRegistry == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.ImageRegistry must not be empty", config)
+	}
+	if config.ResyncPeriod == 0 {
+		return nil, microerror.Maskf(invalidConfigError, "%T.ResyncPeriod must not be empty", config)
 	}
 	if config.WebhookBaseURL == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.WebhookBaseURL not be empty", config)
@@ -93,11 +97,12 @@ func NewApp(config Config) (*App, error) {
 	var appController *controller.Controller
 	{
 		c := controller.Config{
-			InitCtx:   initCtxFunc,
-			K8sClient: config.K8sClient,
-			Logger:    config.Logger,
-			Resources: resources,
-			Selector:  label.AppVersionSelector(config.UniqueApp),
+			InitCtx:      initCtxFunc,
+			K8sClient:    config.K8sClient,
+			Logger:       config.Logger,
+			ResyncPeriod: config.ResyncPeriod,
+			Resources:    resources,
+			Selector:     label.AppVersionSelector(config.UniqueApp),
 			NewRuntimeObjectFunc: func() runtime.Object {
 				return new(v1alpha1.App)
 			},