+++ service/controller/appcatalog/resources.go
@@ -8,7 +8,7 @@ import (
 	"github.com/giantswarm/operatorkit/v4/pkg/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/v4/pkg/resource/wrapper/retryresource"
 
-	"github.com/giantswarm/app-operator/v2/service/controller/appcatalog/resource/appcatalogentry"
+	"github.com/giantswarm/app-operator/v3/service/controller/appcatalog/resource/appcatalogentry"
 )
 
 type appCatalogResourcesConfig struct {