+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Print difference between the current chart and desired chart.
+
 ## [2.8.0] - 2020-12-15
 
 ### Changed