+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Print difference between the current chart and desired chart.
+- Make resync period configurable for use in integration tests.
 
 ## [2.8.0] - 2020-12-15