+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
-- Print difference between the current chart and desired chart.
+- Make resync period configurable for use in integration tests.
 - Pause App CR reconciliation when it has
   `app-operator.giantswarm.io/paused=true` annotation.
+- Print difference between the current chart and desired chart.
 
 ## [2.8.0] - 2020-12-15