+++ CHANGELOG.md
@@ -14,8 +14,10 @@ tenant app CRs.
 
 ### Added
 
-- Print difference between the current chart and desired chart.
 - Make resync period configurable for use in integration tests.
+- Pause App CR reconciliation when it has
+  `app-operator.giantswarm.io/paused=true` annotation.
+- Print difference between the current chart and desired chart.
 
 ## [2.8.0] - 2020-12-15
 
@@ -157,7 +159,7 @@ app CRs. A message is added to the app CR status for the user.
 
 ### Removed
 
-- Don't wait for chart-operator pod since chart CRD is created by the chartcrd resource. 
+- Don't wait for chart-operator pod since chart CRD is created by the chartcrd resource.
 
 ## [2.0.0] - 2020-08-13