+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [3.0.0] - 2021-01-05
+
 ### Changed
 
 - Enable mutating and validating webhooks in app-admission-controller for
@@ -270,7 +272,8 @@ from Helm 2 to Helm 3.
 
 - Flattening operator release structure.
 
-[Unreleased]: https://github.com/giantswarm/app-operator/compare/v2.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-operator/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/giantswarm/app-operator/compare/v2.8.0...v3.0.0
 [2.8.0]: https://github.com/giantswarm/app-operator/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/giantswarm/app-operator/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/giantswarm/app-operator/compare/v2.5.0...v2.6.0